Add ClimberInfo class.
Add a new climbs/index.js file that defines a ClimberInfo
class that stores climber-related information. This is
passed through from the Routes view to the ClimbDropdown
component and contains a map from route ID to ClimbState, a
color, and the climber's name.

Also move the ClimbState enum into this file.