Update Log Cloud Function to adjust for clock skew.
Add a 'now' field to messages sent to the Log Cloud Function
and update the function to use it to adjust records'
timestamps to try to compensate for clock skew between the
client and server.

Also move Go packages into a 'go' subdirectory since it
feels messy to have them scattered across the root of the
repository. The ascenso.go file needs to remain in the root
to make 'gcloud functions deploy' work.