Defer sending log records while offline.
Make the Logger class hold off on scheduling _sendQueued()
calls when the browser reports that the device was offline.
As soon as the device comes online, we schedule sending.

I'm also adding Node types to our TypeScript config. I'm not
completely sure that we'll need this, though, as zero lucks
were given when I tried to mock the navigator.onLine
property in tests.

Finally, I improved comments a bit and switched to Chrome's
style of putting parameter names in pipes (e.g. "Subscribed
to events of type |name|.") to make them easier to read.