Fix toolbar visibility issues. (#74)
Fix an issue where the App view sometimes wouldn't hear
about changes to the logged-in state, resulting in the
toolbar being shown at the login screen or hidden while
logged in. I suspect that I introduced this regression when
I made loggedIn be a computed property rather than a method.