Enable tab-syncing for Firestore persistence. (#81)
I didn't notice it before, but there's an option to tell
Firestore to sync its state between tabs, which allows
persistence a.k.a. offline mode to be used when the site is
open in multiple tabs (which can happen easily after
installing the PWA). Pass this when enabling persistence.

Also make the Routes view display a warning if persistence
is disabled for some reason (ancient browser?).