Add confirmation dialog for "Sign out" nav item. (#83)
Display a confirmation dialog before signing the user out.
The "Sign out" item in the navigation drawer is easy to
accidentally click, and doing so is problematic when the
device doesn't have network connectivity. Also add a divider
between the other items and "Sign out" and write some tests.

Tangentially, switch the titles on dialogs displayed in the
Profile view to use sentence case instead of title case.
This is pretty arbitrary, but I think that "Sign Out" looks
funny and I'd like to be consistent.

Also add a "test:unit:only" script to package.json to make
it easier to run tests from a single file, e.g. "npm run
test:unit:only Toolbar".