Make Profile view search for unused invite codes.
When creating a team, try to find an unused invite code up
to ten times before failing. Before, we'd just try to insert
a document for the new code without checking it first. The
Firestore security rules should reject collisions, but it
seems nicer to handle them gracefully by retrying.

If my math is right, the previously probability of a
collision with 100 existing teams was just 0.0001, but why
not reduce that to 1e-40?

Also fix some broken interpolated error messages.