Improve performance when expanding areas in Routes view.
After the move to Vuetify 2, expanding areas in the Routes
view became super-janky on a Pixel 2 phone running Chrome
77. Setting the 'accordion' attribute on v-expansion-panels
and 'eager' on v-expansion-panel-content seems to improve
this.

Unrelatedly, increase the login timeout in end-to-end tests
from 5 seconds to 15 seconds, since login depends on an
external service and since I just saw it time out.