Support setting grade-based route filters. (#117)
Add support for filtering which routes are displayed in the
Routes view, currently just via a minimum and maximum grade.
A new GradeSlider component is introduced and used in a
dialog in the Routes view. A new RoutesNav component is also
used to place a "Filters" button in the toolbar for opening
the dialog.

Filters are saved to the user doc in Cloud Firestore, but
they aren't reflected in the route listing yet.