Make RouteList visually deemphasize filtered-out routes. (#120)
Update the RouteList component to accept minGrade and
maxGrade parameters. If a range is supplied, routes with
grades outside of it are displayed with reduced opacity.

Update the Routes view to pass the range selected via the
filters dialog.

Also make the 'models' module export a GradeIndexes object
that can be used to look up the index of a given grade in
the Grades array.