Display message about competition start/end time. (#122)
Add startTime and endTime fields to the global/config Cloud
Firestore Document, and update the Routes view to display
the current state of the competition, in the form of a
countdown message.

Also update the deepCopy function in the testutil module to
use lodash's cloneDeep function so that Timestamp objects
are preserved in Firestore docs.

We don't prevent the user from reporting climbs before the
competition starts or after it ends yet. I'm not sure if
that's safe to do or if it risks rejecting climbs that were
recorded earlier without network connectivity.