Upgrade terser-webpack-plugin to resolve vulnerabilities.
Upgrade terser-webpack-plugin from 1.3.0 to 1.4.3 to resolve
serialize-javascript vulnerabilities identified by "npm
audit". For whatever reason, npm also decided to convert a
bunch of packages to be non-dev dependencies. *shrug*

It looks like we still include the affected version of
serialize-javascript due to @vue/cli-service, but updating
that package broke everything the last time I tried to do it
so I'm going to save it for a separate change.