Shorten long UI strings to fit on small screens.
Make various strings in the Profile view shorter so they'll
fit onscreen on iPhone 5/SE, which has a screen width of
just 320 display units. This is particularly a problem in
Spanish.

Also conditionally set the 'small' attribute on the main
buttons in the Profile view for screens with widths up to
400 display units.

Finally, fix an incorrect string in the "Join team" dialog.
It looks like I used the wrong ID here when I finished
translating the Profile view in 3990730e.