Display sync icon in Routes view for pending writes.
Update the Routes view to display a sync icon and the number
of pending updates in the bottom-right corner of the
viewport when there are climb state changes that haven't
been synced with the backend yet.

Also change the way that we import Material icons in
index.html, since the old non-Roboto import doesn't appear
to contain the 'sync' icon. The existing icons in the
Toolbar view's navigation drawer are still rendered
correctly.