Keep UI onscreen while finishing login.
Update the Login view to keep the logo and FirebaseUI auth
container onscreen while finalizing login and waiting for
redirect, and display the spinner in a v-overlay (which
automatically centers it and blocks input).

I wasted a bunch of time trying to figure out why Google
login wasn't working in an incognito window. Apparently
third-party cookies are required, but FirebaseUI seems to
only report this via an ugly div in the auth container
rather than throwing an error or running the signInFailure
callback.