Update to firebase@8.10.0 and firebase-tools@9.21.0.
I initially tried to go to firebase@9.1.3, the current
release, but it redoes the way that module imports work. It
seems straightforward to update our code to use the 'compat'
versions of the modules per the instructions at
https://firebase.google.com/docs/web/modular-upgrade, but
the third-party vuefire package doesn't seem to have been
updated yet (https://github.com/vuejs/vuefire/issues/1128)
so we're stuck at version 8 for now.

Also switch from the deprecated @firebase/testing package to
@firebase/rules-unit-testing. Luckily, it seems to be a
drop-in replacement.

To get all of this working, I also ripped out all of the
code I'd written before to asynchronously import Firebase
modules. It made the code way more complicated for what was
probably only a little bit of improvement in loading time,
and I spent a lot of time trying to figure out how to update
it to work with version 8 of the Firebase API before giving
up. Version 9 finally supports tree-shaking, so hopefully
things will get faster again when we're able to upgrade to
it.