Upgrade to nightwatch@1.7.12 and chromedriver@96.0.0.
Upgrade some code used for end-to-end tests. I think that
chromedriver is supposed to match the Chrome version, but
we were somehow still using chromedriver 80.

And just so it's mentioned somewhere: I was getting errors
like this when running 'npm run test:e2e' and thought it was
due to an incompatibility between Nightwatch and recent
versions of Chrome:

Error: An error occurred while retrieving a new session: "unknown error: Chrome failed to start: crashed."
     at endReadableNT (internal/streams/readable.js:1334:12)
     at processTicksAndRejections (internal/process/task_queues.js:82:21)

After digging a bit and turning on verbose logging by adding
"cli_args: ['--enable-chrome-logs', '--verbose']" to the
'webdriver' section of nightwatch.conf.js, I figured out
that this was actually happening because of my weird setup
(tests were running as a different user which had no
~/.Xauthority file and was thus unable to connect to the X
server -- the 'test:e2e:xvfb' command still worked).