Remove (almost) all uses of the v-t attribute.
The v-t attribute, vue-i18n's version of v-text, apparently
replaces the contents of the element that it's used on, so
it shouldn't be set on components. Otherwise, we get errors
like the following when the 'Leave team' dialog is closed:

  Uncaught DOMException: Failed to execute 'removeChild' on
  'Node': The node to be removed is not a child of this
  node.

I'm still keeping a few uses of v-t in the ClimbDropdown
component, since it's allegedly more performant than $t()
and I haven't noticed any problems there.