Upgrade to Vuetify 2.6.4 (from 2.2.18).
Run 'npm install vuetify@latest' to install vuetify@2.6.4.

That resulted in "SassError: @use is coming soon, but it's
not supported in this version of Dart Sass." when running
'npm run serve', so I also ran 'npm install sass@latest' to
upgrade to sass@1.49.9 (from 1.22.12).

That resulted in a zillion deprecation warnings about
Vuetify using deprecated things in Sass, so then I ran 'npm
install sass@1.25.0' to downgrade to the version that
allegedly introduced @use, which seems to work.

I also had to make some minor tweaks in response to padding
and margin changes in Vuetify.